:truck: move StampGenerator to CreatePage
